requirements-dev.txt
# Requerimientos para Desarrollo
-r requirements.txt
django-debug-toolbar
django-extensions
pytest
pytest-django
